+++ src/main/kotlin/com/teamsparta/buysell/domain/member/controller/MemberController.kt
@@ -6,7 +6,6 @@ import com.teamsparta.buysell.domain.member.dto.request.SignUpRequest
 import com.teamsparta.buysell.domain.member.dto.response.MemberResponse
 import com.teamsparta.buysell.domain.member.service.MemberService
 import com.teamsparta.buysell.domain.member.service.SocialService
-import com.teamsparta.buysell.domain.post.dto.response.PostResponse
 import com.teamsparta.buysell.infra.security.UserPrincipal
 import com.teamsparta.buysell.infra.social.jwt.JwtDto
 import io.swagger.v3.oas.annotations.Operation
@@ -69,30 +68,9 @@ class MemberController(
             .body(memberService.getMyProfile(userPrincipal))
     }
 
-    @GetMapping("/{memberId}/posts")
-    @Operation(summary = "작성한 게시글 조회", description = "작성한 모든 게시글을 조회합니다.")
-    fun getAllPostsMemberId(
-        @PathVariable memberId:Int,
-    ): ResponseEntity<List<PostResponse>> {
-        val posts = memberService.getAllPostByMemberId(memberId)
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(posts)
-    }
-
-    @GetMapping("/likes")
-    @Operation(summary = "내가 찜한 게시글 조회", description = "내가 찜한 게시글을 조회합니다.")
-    fun getLikesByMember(
-        @AuthenticationPrincipal userPrincipal: UserPrincipal,
-    ): ResponseEntity<List<PostResponse>> {
-        val like = memberService.getAllPostByLike(userPrincipal)
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(like)
-    }
 
     @PutMapping("/signout")
-    fun pretendDelete(
+    fun signout(
         @AuthenticationPrincipal userPrincipal: UserPrincipal
     ):ResponseEntity<String>{
         memberService.signOut(userPrincipal)
